requirements.txt
psutil
huggingface_hub
python-dotenv
graphviz
pypdf
python-docx
beautifulsoup4
networkx
numpy
scikit-learn
sentence-transformers
tqdm
streamlit
torch
transformers
accelerate
requests
ninja
pydantic
llama-cpp-python[server]==0.3.1
